Version 3.4.4 It looks great, i guess... NEW: -4 directions to move. SOON: -Nave and enemies have HP. -Another way to shoot. -The nave follows the pointer (disable)

+++ src/Model/Bullet.java
@@ -15,7 +15,7 @@ public class Bullet extends Sprite{
     
     public static final int WIDTH = 5;
     public static final int HEIGHT = 7;
-    private static final int VELOCIDAD = 2;
+    private static final double VELOCIDAD = 2;
     private final int tipo;
     private double vectorX;
     private double vectorY;
