+++ src/Model/Enemy.java
@@ -16,7 +16,7 @@ public class Enemy extends Sprite {
     
     private static final int WIDTH = 18;
     private static final int HEIGHT = 18;
-    private static final int VELOCIDAD = 1;
+    private static final double VELOCIDAD = 1;
     private final int score;
     private static final int MIN = (Enemy.WIDTH/2 + Nave.WIDTH/2);
     private static final int MAX = (Game.WINDOW_WIDTH - Enemy.WIDTH - Nave.WIDTH/2);
